move prepareSnapshot
